+++ cliquet_fxa/views/relier.py
@@ -71,7 +71,7 @@ def fxa_oauth_login(request):
                 '&client_id={client_id}&state={state}&scope={scope}')
     form_url = form_url.format(oauth_uri=fxa_conf(request, 'oauth_uri'),
                                client_id=fxa_conf(request, 'client_id'),
-                               scope=fxa_conf(request, 'scope'),
+                               scope=fxa_conf(request, 'requested_scope'),
                                state=state)
     request.response.status_code = 302
     request.response.headers['Location'] = form_url